# Conflicts:
#	gieFrontend/src/main/java/org/giefront/EditDeleteCellEnt.java
#	gieFrontend/src/main/java/org/giefront/EntropriseModificationController.java
#	gieFrontend/src/main/java/org/giefront/FourniseurEntoContro.java
#	gieFrontend/src/main/java/org/giefront/IEntropriseModificationController.java
#	gieFrontend/src/main/java/org/giefront/Service/EntrepriseService.java
#	gieFrontend/src/main/java/org/giefront/TestFront.java
#	gieFrontend/src/main/java/org/giefront/controller/AdvancedSearchE.java
#	gieFrontend/src/main/java/org/giefront/controller/EntrepriseModifController.java
#	gieFrontend/src/main/resources/org/giefront/EntrepriseModification.fxml
#	gieFrontend/src/main/resources/org/giefront/PersonneModification.fxml

+++ gieBackend/src/main/java/org/gieback/Controller/PersonneController.java
@@ -101,7 +101,7 @@ public class PersonneController {
 
     @DELETE
     @Path("DeletePersonne/{id}")
-    public Response deleteById(@PathParam("id") int id) {
+    public Response deleteById(@PathParam("id") Long id) {
         personneService.deleteById(id);
         return Response.noContent().build();
     }
